+++ src/main/frontend/src/components/UI/SearchResults.jsx
@@ -1,67 +1,18 @@
-import { useState, useRef, useEffect } from "react";
+import { useRef } from "react";
 import SearchBar from "./SearchBar";
 import { motion } from "framer-motion";
-import { useMutation, useQuery } from "@tanstack/react-query";
-import { fetchSpots } from "../../util/http";
-import { useParams } from "react-router-dom";
-import ErrorPage from "./Error/ErrorPage";
+
+import { useSearch } from "../../hooks/useSearch";
 function SearchResults({ apiMode, searchId, items, children }) {
   const lastTerm = useRef();
-  const params = useParams();
-  const [searchTerm, setSearchTerm] = useState("");
 
-  const searchResults = items.filter((item) =>
-    JSON.stringify(item).toLowerCase().includes(searchTerm.toLowerCase())
-  );
-  const { data, mutate, isPending, isError, error } = useMutation({
-    mutationKey: ["spots", { term: searchTerm }],
-    mutationFn: () => fetchSpots(searchTerm, params.city),
+  const { content, onSubmitHandler, handleChange } = useSearch({
+    items,
+    apiMode,
+    lastTerm,
+    children,
   });
 
-  function handleChange(event) {
-    if (lastTerm.current) {
-      clearTimeout();
-    }
-    lastTerm.current = setTimeout(() => {
-      lastTerm.current = null;
-      setSearchTerm(event.target.value);
-    }, 1000);
-  }
-  function onSubmitHandler(event) {
-    event.preventDefault();
-    setSearchTerm(event.target.searchTerm.value);
-    mutate();
-  }
-  let content;
-
-  if (!apiMode) {
-    content = searchResults.map((item, isClicked) => (
-      <motion.li layout key={item.id ? item.id : item.citySeq}>
-        {children(item, isClicked)}
-      </motion.li>
-    ));
-  } else {
-    if (isPending) {
-      content = <p>데이터를 가져오는 중입니다...</p>;
-    }
-    if (isError) {
-      content = (
-        <ErrorPage
-          title="에러발생"
-          message={error.info?.message || "여행지를 가져오는 데 실패했습니다."}
-        />
-      );
-    }
-    if (data) {
-      content = data.places.map((item, isClicked) => (
-        <motion.li layout key={item.id ? item.id : item.citySeq}>
-          {children(item, isClicked)}
-        </motion.li>
-      ));
-    } else {
-      content = <p>검색 결과가 없습니다.</p>;
-    }
-  }
   return (
     <div className="flex flex-col items-center">
       <SearchBar
