+++ src/main/frontend/src/pages/SelectCityPage.jsx
@@ -16,7 +16,6 @@ import BottomSheet from "../components/UI/Bottomsheet/BottomSheet";
 import { useQuery } from "@tanstack/react-query";
 import { fetchCities } from "../util/http";
 import ErrorPage from "../components/UI/Error/ErrorPage";
-import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 
 const SelectCityPage = () => {
@@ -50,7 +49,7 @@ const SelectCityPage = () => {
   };
   // 나라 목록 컨텐츠
   let countryContent = (
-    <SearchResults items={countryData}>
+    <SearchResults items={countryData} apiMode={false}>
       {(item, index) => {
         if (index < 10) {
           return (
